Corrected win32 keyboard listener
The previous implementation called ToUnicodeEx in the hook procedure.
This function is documented to modify internal kernel structures, and
doing this in a hook procedure causes data races, which corrupted
especially dead keys.

The new implementation performs an initial layout extraction and uses
that together with the current modifier state to translate events.